imp: update ios with new models
