+++ android/src/main/kotlin/dev/steenbakker/flutter_ble_peripheral/handlers/StateChangedHandler.kt
@@ -20,7 +20,6 @@ class StateChangedHandler(flutterPluginBinding: FlutterPlugin.FlutterPluginBindi
     )
 
     init {
-
         eventChannel.setStreamHandler(this)
     }
